requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.0.3
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
selenium==4.11.2
wordcloud==1.9.2
matplotlib==3.7.2
konlpy==0.6.0
newspaper3k==0.2.8
flask-cors==4.0.0
